Fixing index information retrieval
- Storing doubles instead of floats
- Fixing integer values parsing

--HG--
branch : dev
extra : transplant_source : %EB%7E%27%E7%3B%2B%CEvA%AE%90%3FQ%BE%9Bw%9C%3FXA

+++ src/Orchard/Indexing/IIndexDocument.cs
@@ -10,7 +10,7 @@ namespace Orchard.Indexing {
         IDocumentIndex Add(string name, DateTime value);
         IDocumentIndex Add(string name, int value);
         IDocumentIndex Add(string name, bool value);
-        IDocumentIndex Add(string name, float value);
+        IDocumentIndex Add(string name, double value);
 
         /// <summary>
         /// Stores the original value to the index.
